[core] Remove branch merge from catalog API
