+++ paimon-core/src/main/java/org/apache/paimon/catalog/DelegateCatalog.java
@@ -264,12 +264,6 @@ public abstract class DelegateCatalog implements Catalog {
     }
 
     @Override
-    public void mergeBranch(Identifier identifier, String sourceBranch, String targetBranch)
-            throws BranchNotExistException {
-        wrapped.mergeBranch(identifier, sourceBranch, targetBranch);
-    }
-
-    @Override
     public List<String> listBranches(Identifier identifier) throws TableNotExistException {
         return wrapped.listBranches(identifier);
     }
